feat: 유저 유효성 검증, ResponseEntity 반환

+++ src/main/java/com/nam20/news_invest/service/UserService.java
@@ -1,10 +1,13 @@
 package com.nam20.news_invest.service;
 
 import com.nam20.news_invest.entity.User;
+import com.nam20.news_invest.exception.ResourceNotFoundException;
 import com.nam20.news_invest.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -20,18 +23,42 @@ public class UserService {
     }
 
     public User retrieveUser(Long id) {
-        return userRepository.findById(id).orElse(null);
+        Optional<User> optionalUser = userRepository.findById(id);
+
+        if (optionalUser.isEmpty())
+            throw new ResourceNotFoundException("id: " + id);
+
+        return optionalUser.get();
     }
 
-    public User saveUser(User user) {
+    public User createUser(User user) {
+        user.setCreatedAt(LocalDateTime.now());
+        user.setUpdatedAt(LocalDateTime.now());
         return userRepository.save(user);
     }
 
     public void deleteUser(Long id) {
+        Optional<User> optionalUser = userRepository.findById(id);
+
+        if (optionalUser.isEmpty())
+            throw new ResourceNotFoundException("id: " + id);
+
         userRepository.deleteById(id);
     }
 
-    public User updateUser(User user) {
-        return userRepository.save(user);
+    public User updateUser(Long id, User user) {
+        Optional<User> optionalUser = userRepository.findById(id);
+
+        if (optionalUser.isEmpty())
+            throw new ResourceNotFoundException("id: " + id);
+
+        User existingUser = optionalUser.get();
+
+        existingUser.setName(user.getName());
+        existingUser.setEmail(user.getEmail());
+        existingUser.setPassword(user.getPassword());
+        existingUser.setUpdatedAt(LocalDateTime.now());
+
+        return userRepository.save(existingUser);
     }
 }
